refactor(widget-plugin-rollup-build-ts): move print and cleanup to separate files
